+++ sample/index.js
@@ -12,7 +12,10 @@ function main() {
 
     const deepgram = new Deepgram({ apiKey: config.deepgramApiKey, apiSecret: config.deepgramApiSecret });
 
-    deepgram.transcribe(config.urlToFile, { punctuate: true })
+    /**
+     * Transcribing a file 
+     */
+    deepgram.transcription.preRecorded({ url: config.urlToFile }, { punctuate: true })
       .then((result) => {
         console.log(result.results.channels[0]);
       })
@@ -20,21 +23,35 @@ function main() {
         console.log(err);
       });
 
+    /**
+     * Create & delete projects and API keys
+     */
     let key = 'Xwr5JGdWhUciiNLZ';
-    deepgram.keys.create('test')
-      .then((result) => {
-        key = result.key
+    deepgram.projects.create('testproject')
+      .then((project) => {
 
-        if (key) {
-          deepgram.keys.delete(key)
-            .then((result) => {
-              console.log(`Successfully deleted key: ${JSON.stringify(result)}`);
-            })
-            .catch((err) => {
-              console.log(err);
-            });
-        }
+        deepgram.keys.create(project.project_uuid, key, ['account:read'])
+          .then((key) => {
+            if (key) {
+              deepgram.keys.delete(project.project_uuid, key.key)
+                .then((result) => {
+                  deepgram.projects.delete(project.project_uuid)
+                    .then((r) => {
 
+                      console.log(`Successfully deleted key: ${JSON.stringify(result)}`);
+                    })
+                    .catch((err) => {
+                      console.log(err);
+                    });
+                })
+                .catch((err) => {
+                  console.log(err);
+                });
+            }
+          })
+          .catch((err) => {
+            console.log(err);
+          });
       })
       .catch((err) => {
         console.log(err);