+++ src/types/apiKeyResponse.ts
@@ -1,5 +1,0 @@
-import { Key } from "node:readline";
-
-export type ApiKeyResponse = {
-  keys: Array<Key>;
-};