+++ src/types/keyResponse.ts
@@ -1,0 +1,5 @@
+import { Key } from "./key";
+
+export type KeyResponse = {
+  keys: Array<Key>;
+};