+++ src/types/project.ts
@@ -1,0 +1,13 @@
+/**
+ * Deepgram project
+ */
+export type Project = {
+  /**
+   * Unique identifier of the project
+   */
+  project_uuid: string;
+  /**
+   * User provided name of the project
+   */
+  name: string;
+};