+++ src/types/projectResponse.ts
@@ -1,0 +1,5 @@
+import { Project } from "./project";
+
+export type ProjectResponse = {
+  projects: Array<Project>;
+};